Round-6 fixes: bam leverage, bs=:re semantics, type-stability, calibration tests
Correctness:
- bam computed hat diagonals without the weight factor at both sites, so
  leverage() and cooksdistance() were silently wrong for every non-Gaussian
  or weighted bam fit (max leverage 161 where it must be <= 1; max Cook's
  1.0e21). Fixed structurally: a shared pirls_finalize now owns EDF, the
  weighted leverage identity and R, so scasm and pirls_extended inherit it
  without edits, and bam's halving moved onto pirls_halve! (valid-mu
  enforcement arrives with it). Bam fits are bit-identical.
- bs=:re now follows mgcv's rule (model.matrix(~ v1:v2:...:vk - 1)):
  factors expand to one column per level, numerics enter linearly, and the
  rule is symmetric in the variables. Verified 8/8 cases against mgcv,
  including two previously erroring. This fixes s(x) on a continuous
  covariate silently building an n x n design (382 MiB at n=5000). Integer
  grouping columns now warn, since they are treated as numeric.
- Unseen factor levels at predict no longer silently return the reference
  level. Checked against mgcv 1.9.4 rather than assumed: mgcv warns and
  contributes nothing for such rows (it does not error, and does not fall
  back to the reference), so both the parametric and factor-by paths now
  warn and contribute zero.
- predict(type=<unrecognized>) errored only on the multi-parameter path;
  all paths now reject it. k > n_unique now errors for cr, cs, cc and gp
  as it already did for the thin-plate family.

Performance (the round-5 type-inference bug class, two more instances):
- basis_gp: a Float64(get(xt, ...)) inferring Any poisoned an n x k loop.
  200,908 -> 146 allocations.
- _bspline_basis allocated its buffers inside the per-row loop, affecting
  ps/cps/bs/ad, every SCAM and SCASM basis, ps-margin tensors and nested
  bases, in both fitting and prediction. 100,003 -> 7 allocations.
  Bit-identical: 31/31 references, max difference exactly 0.0.
- penalty_edf's leverage computation is chunked: 7.1 -> 3.0 MiB, exact.

Inference and documentation:
- AIC now matches mgcv's convention (previous commit) and summary(m)
  prints an mgcv-style table with significance codes and a criterion
  footer, instead of Base's type-name fallback.
- The README Quick Start now runs verbatim; it died on coef(m) because
  using StatsAPI after using GAM leaves the verbs ambiguous.
- Two over-warnings corrected against measurement: SCAM interval coverage
  is 0.948 (nominal 0.95), not the shortfall a bootstrap ratio implied;
  and the testStat simplification changes Ref.df and F but not test size
  (0.037 vs mgcv 0.035 at alpha=0.05).
- Documenter build fixed: it was failing outright because api.md exceeded
  a size limit; now three pages, exit 0, warnings 13 -> 7.

Tests: new calibration suite (interval coverage, null test size, SCAM
coverage, select= shrinkage) verified to catch a Vp scaling error in both
directions; leverage/EDF identity across fitters x families (failed on
bam before the fix); allocation-regression bounds; summary() coverage.

Full suite with live R: 4558/4558.

+++ src/basis_extra.jl
@@ -520,7 +520,7 @@ function _smooth_construct(::SoapFilm, spec::SmoothSpec, data, user_knots)
 
     # ── Unpack inputs ────────────────────────────────────────────────────
     bnd = Matrix{Float64}[Float64.(b) for b in spec.xt[:bnd]]
-    nmax = Int(get(spec.xt, :nmax, 200))
+    nmax = Int(get(spec.xt, :nmax, 200))::Int
 
     x = Float64.(Tables.getcolumn(data, spec.term_vars[1]))
     y = Float64.(Tables.getcolumn(data, spec.term_vars[2]))
